:recycle: refactor all code to change gomoko to ft_gomoku
