+++ ft_gomoku/data_structure/SettingsStruct.py
@@ -1,4 +1,4 @@
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 import os
 import json
