+++ ft_gomoku/engine/sound_control.py
@@ -7,6 +7,6 @@ def load_music(music_file: str, repeat: int = -1):
     :param music_file: music file name.
     :param repeat: how many times to repeat. -1 for infinite.
     """
-    path = os.path.join('gomoku', 'assets', 'music', music_file)
+    path = os.path.join('ft_gomoku', 'assets', 'music', music_file)
     pygame.mixer.music.load(path)
     pygame.mixer.music.play(repeat)
