+++ ft_gomoku/game/__init__.py
@@ -1,4 +1,2 @@
-from .engine import Engine, sound_control
 from .screen import main_menu, game_screen
 from .screen import Particle, stars_effect
-from .engine import get_image
