+++ ft_gomoku/game/screen/components/maximize_button.py
@@ -1,10 +1,10 @@
 import pygame
-from gomoku.game.engine import Engine
+from ft_gomoku.engine import Engine
 
 
 def maximize_button(engine: Engine):
 	"""Maximize button"""
-	image_maximize = pygame.image.load('gomoku/assets/img/maximize.png')
+	image_maximize = pygame.image.load('ft_gomoku/assets/img/maximize.png')
 	image_maximize = pygame.transform.scale(image_maximize, (50, 50))
 	image_maximize_rect = image_maximize.get_rect()
 	image_maximize_rect.bottomright = (engine.get_window_size()[0] - 30, engine.get_window_size()[1] - 120)
