+++ ft_gomoku/game/screen/components/mute_button.py
@@ -1,11 +1,11 @@
 import pygame
-from gomoku.game.engine import Engine
+from ft_gomoku.engine import Engine
 
 
 def mute_button(engine: Engine):
 	"""Mute button"""
-	image_mute = pygame.image.load('gomoku/assets/img/volume-mute.png')
-	image_unmute = pygame.image.load('gomoku/assets/img/volume.png')
+	image_mute = pygame.image.load('ft_gomoku/assets/img/volume-mute.png')
+	image_unmute = pygame.image.load('ft_gomoku/assets/img/volume.png')
 	image_mute = pygame.transform.scale(image_mute, (50, 50))
 	image_unmute = pygame.transform.scale(image_unmute, (50, 50))
 	image_mute_rect = image_mute.get_rect()
