--- gomoku/game_rules/__init__.py
@@ -1,2 +1,0 @@
-from .rule_status import RuleStatus
-from .five_to_win_rule import five_to_win
