--- gomoku/game_rules/rule_status.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class RuleStatus(Enum):
-	NO = -1
-	OK = 0
-	WIN = 1
